import checkandget functionality
